rewrite input component to TSX
